chore(refactor-models): move models into logically organized modules
