+++ fence/models/_base.py
@@ -1,0 +1,5 @@
+"""
+Base model
+"""
+from sqlalchemy.ext.declarative import declarative_base
+Base = declarative_base()
